Refactor rust linters into separate modules
