+++ test/lint/test_runner/src/linters.rs
@@ -1,0 +1,9 @@
+// Copyright (c) The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or https://opensource.org/license/mit/.
+
+pub mod doc;
+pub mod includes;
+pub mod std_filesystem;
+pub mod subtree;
+pub mod whitespace;
