+++ test/lint/test_runner/src/linters/whitespace.rs
@@ -1,0 +1,54 @@
+// Copyright (c) The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or https://opensource.org/license/mit/.
+
+use crate::{exclude, utils, LintResult};
+
+pub fn trailing_whitespace() -> LintResult {
+    let trailing_space = utils::git()
+        .args(["grep", "-I", "--line-number", "\\s$", "--"])
+        .args(exclude::get_pathspecs_exclude_whitespace())
+        .status()
+        .expect("command error")
+        .success();
+    if trailing_space {
+        Err(r#"
+^^^
+Trailing whitespace is problematic, because git may warn about it, or editors may remove it by
+default, forcing developers in the future to either undo the changes manually or spend time on
+review.
+
+Thus, it is best to remove the trailing space now.
+
+Please add any false positives, such as subtrees, Windows-related files, patch files, or externally
+sourced files to the exclude list.
+            "#
+        .to_string())
+    } else {
+        Ok(())
+    }
+}
+
+pub fn tabs_whitespace() -> LintResult {
+    let tabs = utils::git()
+        .args(["grep", "-I", "--line-number", "--perl-regexp", "^\\t", "--"])
+        .args(["*.cpp", "*.h", "*.md", "*.py", "*.sh"])
+        .args(exclude::get_pathspecs_exclude_whitespace())
+        .status()
+        .expect("command error")
+        .success();
+    if tabs {
+        Err(r#"
+^^^
+Use of tabs in this codebase is problematic, because existing code uses spaces and tabs will cause
+display issues and conflict with editor settings.
+
+Please remove the tabs.
+
+Please add any false positives, such as subtrees, or externally sourced files to the exclude list.
+            "#
+        .to_string())
+    } else {
+        Ok(())
+    }
+}
